refactor: delegated files to its correct folders
