+++ app/src/main/java/com/example/valorant_app/ui/pages/agent/card/compose/FlagViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.valorant_app.ui.pages.agent.card
+package com.example.valorant_app.ui.pages.agent.card.compose
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
